progress: add panic for multiwriter cycles
Previously, if this case occured, we would deadlock, so as to make this
case more easily identifiable and debuggable, we should panic instead.
Potentially, we could also just *skip* this case, but given that this is
an internal logical error that should not occur, it would be better to
identify this.

Signed-off-by: Justin Chadwell <me@jedevc.com>